reuse Markers component
